+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.28.0
+
+### Breaking Changes
+
+#### Cloud config
+
+- `deploymentId` and `encryptedDeploymentId` removed from **uploadAdapter** options
+- `scope` added to **uploadAdapter** options
+
+### Added
+
+#### resolve runtime
+
+- `clientIp` in request object
+
+
 ## 0.27.0
 
 ### Breaking Changes